requirements.txt
Django~=3.2.4
psycopg2-binary~=2.9
djangorestframework~=3.12.4
django-cors-headers~=3.7.0
gunicorn~=20.1.0